Update from original - main

+++ test/test_runner.sh
@@ -3,13 +3,15 @@
 # Get current test file name
 TESTFILE_NAME=${PGAPPNAME##pg_regress/}
 
-if [ "$PARALLEL" -eq 0 ]; then
+if [ "$PARALLEL" -eq 1 ]; then
+    # parallel tests all run in the same database
+    TEST_CASE_DB="ldb_parallel"
+elif [ "$MISC" -eq 1 ]; then
+    TEST_CASE_DB="ldb_misc"
+else
     # Set different name for each test database
     # As pg_regress does not support cleaning db after each test
     TEST_CASE_DB="ldb_test_${TESTFILE_NAME}"
-else
-    # parallel tests all run in the same database
-    TEST_CASE_DB="ldb_parallel"
 fi
 
 # Set database user
@@ -52,7 +54,7 @@ function run_regression_test {
 cd sql/
 
 # install lantern extension
-if [[ "$PARALLEL" -eq 0 || "$TESTFILE_NAME" == "begin" ]]; then
+if [[ ("$PARALLEL" -eq 0  &&  "$MISC" -eq 0) || "$TESTFILE_NAME" == "begin" ]]; then
      psql "$@" -U ${DB_USER} -d postgres -v ECHO=none -q -c "DROP DATABASE IF EXISTS ${TEST_CASE_DB};" 2>/dev/null
      psql "$@" -U ${DB_USER} -d postgres -v ECHO=none -q -c "CREATE DATABASE ${TEST_CASE_DB};" 2>/dev/null
 fi
@@ -68,11 +70,28 @@ then
      # psql "$@" -U ${DB_USER} -d ${TEST_CASE_DB} -q -c "SELECT * FROM pg_extension_update_paths('lantern');" 2>/dev/null
      # install the old version of the extension and sanity-check that all tests pass
      psql "$@" -U ${DB_USER} -d ${TEST_CASE_DB} -v ECHO=none -q -c "SET client_min_messages=error; CREATE EXTENSION IF NOT EXISTS lantern VERSION '$UPDATE_FROM';"
-     psql "$@" -U ${DB_USER} -d ${TEST_CASE_DB} -v ECHO=none -q -f utils/common.sql 2>/dev/null
      # upgrade to the new version of the extension and make sure that all existing tests still pass
      # todo:: this approach currently is broken for pgvector-compat related upgrade scripts as that regression test drops
      # and recreates the extension so whatever we do here is ignored
-     psql "$@" -U ${DB_USER} -d ${TEST_CASE_DB} -v ECHO=none -q -c "SET client_min_messages=error; ALTER EXTENSION lantern UPDATE TO '$UPDATE_TO';"
+     # parallel tests run into issues when multiple instances of the runner simultaneously reindex, we need to track when
+     # this occurs and make the process conditional on it
+     LOCKFILE="/tmp/ldb_update.lock"
+     FINISHEDFILE="/tmp/ldb_update_finished"
+
+     (
+         if flock -xn 200; then
+             if [ ! -f "$FINISHEDFILE" ]; then
+                 psql "$@" -U ${DB_USER} -d ${TEST_CASE_DB} -v ECHO=none -q -f utils/common.sql 2>/dev/null
+                 psql "$@" -U ${DB_USER} -d ${TEST_CASE_DB} -v ECHO=none -q -c "SET client_min_messages=error; ALTER EXTENSION lantern UPDATE TO '$UPDATE_TO';" 2>/dev/null
+                 touch $FINISHEDFILE
+             fi
+         fi
+     ) 200>"$LOCKFILE"
+
+     while [ ! -f "$FINISHEDFILE" ]; do
+         sleep 1
+     done
+
      run_regression_test $@
 else
